## 자판기 프로젝트 구성도 issue resolve - #15 ### drink-api > - 음료수를 진열,삭제,추가 하는 api가 들어가는 모듈 > - 기능에 큰 도메인 별로 패키지를 분리하여 설계하였음 > - drink-api는 따로 was로 올라가지 않는다. ### application > - url에 따른 타입리프 템플릿을 보여주는 view와 관련된 모듈 > - resource 파일들이 위치할 예정 , Thymeleaf 엔진을 활용 > - 서버는 application was 하나로 올라간다. ### docker > - docker 관련된 파일을 모을 예정, compose나 dockerfile >   - 현재는 mongodb 관련 Dockerfile 추가
