--- api/build.gradle
@@ -1,10 +1,0 @@
-bootJar {
-    enabled = true
-    archiveName 'DRINK.API.jar'
-}
-
-version = 'v01'
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-}
