--- api/src/main/java/com/example/drink/dto/DrinkDto.java
@@ -1,22 +1,0 @@
-package com.example.drink.dto;
-
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-@ToString
-public class DrinkDto {
-
-    private long drinkId;
-    private String drinkName;
-    private int price;
-    private int quantity_count;
-
-    /**
-     * 음료 구매 여부
-     * amount: 보유 금액
-     */
-    public Boolean enablePurchase(int amount) {
-        return amount >= price;
-    }
-}
