--- api/src/main/java/com/example/drink/dto/OrderDto.java
@@ -1,12 +1,0 @@
-package com.example.drink.dto;
-
-import lombok.Data;
-import lombok.ToString;
-
-@Data
-@ToString
-public class OrderDto {
-
-    private String drinkName;
-    private int drinkPrice;
-}
