+++ application/src/main/java/com/DrinkApplication.java
@@ -1,0 +1,13 @@
+package com;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DrinkApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DrinkApplication.class, args);
+    }
+
+}
