+++ application/src/main/resources/application.yml
@@ -1,0 +1,19 @@
+drink:
+  product:
+    list: 콜라,사이다,식혜
+spring:
+  data:
+    mongodb:
+      uri: mongodb://localhost:27017
+      database: vending-machine
+  mvc:
+    static-path-pattern: /static/**
+  thymeleaf:
+    cache: false
+    check-template-location: true
+    prefix: classpath:static/templates/
+    suffix: .html
+
+  devtools:
+    livereload:
+      enabled: true
