+++ application/src/main/resources/static/templates/mainframe.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>자판기</title>
+</head>
+<body>
+<h1 th:text="${name}">Name</h1>
+
+<p>자판기</p>
+
+<div id="selectDrink">
+    <label><input class="drinkCheckBox" type="checkbox" name="coke" value="콜라">콜라 2000원</label>
+    <label><input class="drinkCheckBox" type="checkbox" name="sikhye" value="식혜">식혜 3000원</label>
+    <label><input class="drinkCheckBox" type="checkbox" name="sprite" value="사이다">사이다 1000원</label>
+</div>
+<br>
+<br>
+<br>
+<p>지출할 금액</p><input id="amount" type="text" value="" />
+<br>
+<button id="itemBuyCall">구입</button>
+
+<br>
+<br>
+<br>
+
+<div id="receipt">
+
+</div>
+
+</body>
+<script type="text/javascript" src="/js/mainframe.js"></script>
+</html>
