+++ docker/mongo/Dockerfile
@@ -1,0 +1,9 @@
+FROM mongo:latest
+
+ENV MONGO_INITDB_ROOT_USERNAME seongheon
+ENV MONGO_INITDB_ROOT_PASSWORD qwer1234
+ENV MONGO_INITDB_DATABASE vending-machine
+
+COPY init/init_data_mongo.js /docker-entrypoint-initdb.d/
+
+EXPOSE 27017
