+++ drink-api/build.gradle
@@ -1,0 +1,13 @@
+
+jar {
+    enabled = true
+    archiveName 'DRINK.API.jar'
+}
+
+version = 'v01'
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
+    testImplementation 'org.testcontainers:mongodb:1.17.2'
+}
