+++ drink-api/src/main/java/com/example/display/dto/DisplayDrinkDto.java
@@ -1,0 +1,4 @@
+package com.example.display.dto;
+
+public class DisplayDrinkDto {
+}
