+++ drink-api/src/main/java/com/example/mongo/config/MongoConfigurer.java
@@ -1,0 +1,30 @@
+package com.example.mongo.config;
+
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.MongoClients;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.mongo.MongoProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+@Configuration
+@EnableMongoRepositories(basePackages = {"com.example"})
+public class MongoConfigurer {
+
+    private MongoProperties props = null;
+
+    @Value("${spring.data.mongodb.uri}")
+    private String mongoUrl;
+
+    @Bean
+    public MongoClient mongoClient() {
+        return MongoClients.create(mongoUrl);
+    }
+
+    @Bean
+    public MongoTemplate mongoTemplate() {
+        return new MongoTemplate(mongoClient(), "vending-machine");
+    }
+}
