--- mongodb/Dockerfile
@@ -1,1 +1,0 @@
-mongodb 관련 도커 파일
