--- mongodb/build.gradle
@@ -1,11 +1,0 @@
-bootJar {
-    enabled = false
-}
-jar {
-    enabled = true
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
-    testImplementation 'org.testcontainers:mongodb:1.17.1'
-}
