--- mongodb/init_data_mongo/vending-machine/drink.json
@@ -1,4 +1,0 @@
-[{ "_id" : 0 , "name": "콜라" , "price": 2000 , "quantity" : 5},
-{ "_id": 1 , "name" : "식혜" , "price": 3000 , "quantity" : 5},
-{ "_id": 2 , "name": "사이다" , "price": 1000 , "quantity" : 5}]
-
