--- mongodb/src/main/java/com/example/drink/controller/MainController.java
@@ -1,1 +1,0 @@
-package com.example.drink.controller;
