--- mongodb/src/main/java/com/example/drink/model/Drink.java
@@ -1,19 +1,0 @@
-package com.example.drink.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.data.mongodb.core.mapping.MongoId;
-
-@Document
-@Data
-@AllArgsConstructor
-public class Drink {
-
-    @MongoId
-    ObjectId id;
-    String name;
-    long price;
-    long quantity;
-}
