--- mongodb/src/main/java/com/example/drink/repo/DrinkRepository.java
@@ -1,15 +1,0 @@
-package com.example.drink.repo;
-
-import com.example.drink.model.Drink;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface DrinkRepository extends MongoRepository<Drink,String> {
-
-    Optional<Drink> findByName(String name);
-
-
-}
