--- web/Dockerfile
@@ -1,1 +1,0 @@
-web 관련 도커파일
