--- web/src/main/java/com/PracticeApplication.java
@@ -1,13 +1,0 @@
-package com;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PracticeApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PracticeApplication.class, args);
-    }
-
-}
